Refac: split server-side code
